refactor: restructured project and move revalidation frequency to env variable
