+++ lib/definitions.ts
