+++ lib/utils.ts
@@ -1,0 +1,2 @@
+export const isNumber = (s?: string): boolean =>
+  s != null && s !== "" && !isNaN(Number(s.toString()));
