+++ package.json
@@ -26,6 +26,7 @@
     "eslint": "^8",
     "eslint-config-next": "14.0.4",
     "eslint-plugin-next": "^0.0.0",
+    "json-server": "^1.0.0-alpha.23",
     "postcss": "^8",
     "prettier": "3.1.1",
     "tailwindcss": "^3.3.0",
